Update armor.json
Adjusted large back holster to be actually large enough for heavy weapons. This is doom guys armor. He uses big stupid guns. The back holster is now big enough to actually fit a BFG big fucking gun ( matched max weight, volume, and length to the specs of the BFG 2041-2)

Similarly, the small pockets, while not flagged as holsters (so you can fill them with drugs instead of guns, as per the description of carrying multiple guns as well as stim packs and stuff), are dreadfully slow at 80 moves to retrieve a small arm. There are regular holsters that have as little as 20 or 30 moves, but these are not gun specific pockets so I didn't bring them that low, but to a much improved 40 moves as a sort of hybrid pocket/holster.