Delete aframebackpack.json
ALL the recipes defined in this file are duplicates, so instead of just deleting one ill just rename the file too .txt and leave it as is. idk why theres 2 copies of IDs and such with slightly different stats, but leaving it as a .txt makes it easy for anyone to change which one is the functioning one while removing errors upon using the mod.